refactor(database): rename filename to find the models by table name easily
