+++ app/database/achievement.py
@@ -24,7 +24,7 @@ from sqlmodel import (
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 if TYPE_CHECKING:
-    from .lazer_user import User
+    from .user import User
 
 
 class UserAchievementBase(SQLModel, UTCBaseModel):
