+++ app/database/auth.py
@@ -19,7 +19,7 @@ from sqlmodel import (
 )
 
 if TYPE_CHECKING:
-    from .lazer_user import User
+    from .user import User
 
 
 class OAuthToken(UTCBaseModel, SQLModel, table=True):
