+++ app/database/daily_challenge.py
@@ -17,7 +17,7 @@ from sqlmodel import (
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 if TYPE_CHECKING:
-    from .lazer_user import User
+    from .user import User
 
 
 class DailyChallengeStatsBase(SQLModel, UTCBaseModel):
