+++ app/database/score_token.py
@@ -5,7 +5,7 @@ from app.models.score import GameMode
 from app.utils import utcnow
 
 from .beatmap import Beatmap
-from .lazer_user import User
+from .user import User
 
 from sqlalchemy import Column, DateTime, Index
 from sqlalchemy.orm import Mapped
