+++ app/router/private/cover.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import hashlib
 
-from app.database.lazer_user import User, UserProfileCover
+from app.database.user import User, UserProfileCover
 from app.dependencies.database import Database
 from app.dependencies.storage import get_storage_service
 from app.dependencies.user import get_client_user
