+++ app/router/v1/user.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from datetime import datetime
 from typing import Literal
 
-from app.database.lazer_user import User
 from app.database.statistics import UserStatistics, UserStatisticsResp
+from app.database.user import User
 from app.dependencies.database import Database, get_redis
 from app.log import logger
 from app.models.score import GameMode
