+++ app/router/v2/relationship.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from app.database import Relationship, RelationshipResp, RelationshipType, User
-from app.database.lazer_user import UserResp
+from app.database.user import UserResp
 from app.dependencies.api_version import APIVersion
 from app.dependencies.database import Database
 from app.dependencies.user import get_client_user, get_current_user
