+++ app/router/v2/tags.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from app.database.beatmap import Beatmap
 from app.database.beatmap_tags import BeatmapTagVote
-from app.database.lazer_user import User
 from app.database.score import Score
+from app.database.user import User
 from app.dependencies.database import get_db
 from app.dependencies.user import get_client_user
 from app.models.score import Rank
